trade_exception_predictor/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
azure-ai-ml==1.32.0
azure-common==1.1.28
azure-core==1.39.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.3
azure-mgmt-core==1.6.0
azure-monitor-opentelemetry==1.8.2
azure-monitor-opentelemetry-exporter==1.0.0b45
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.24.0
azureml-mlflow==1.60.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.105.0
docker==7.1.0
fastapi==0.136.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.47
google-auth==2.49.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
h11==0.16.0
huey==2.6.0
idna==3.13
importlib_metadata==8.7.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
Mako==1.3.11
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.9
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
msal==1.36.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.1
numpy==2.4.4
oauthlib==3.3.1
opentelemetry-api==1.41.1
opentelemetry-instrumentation==0.62b1
opentelemetry-instrumentation-asgi==0.62b1
opentelemetry-instrumentation-dbapi==0.62b1
opentelemetry-instrumentation-django==0.62b1
opentelemetry-instrumentation-fastapi==0.62b1
opentelemetry-instrumentation-flask==0.62b1
opentelemetry-instrumentation-psycopg2==0.62b1
opentelemetry-instrumentation-requests==0.62b1
opentelemetry-instrumentation-urllib==0.62b1
opentelemetry-instrumentation-urllib3==0.62b1
opentelemetry-instrumentation-wsgi==0.62b1
opentelemetry-proto==1.41.1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
opentelemetry-util-http==0.62b1
packaging==26.2
pandas==2.3.3
pillow==12.2.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydash==8.0.6
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
skops==0.14.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
strictyaml==1.7.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.8
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.1
